refactor(web): move the guided face-cleanup dashboard to /admin/face-cleanup/scan
Frees /admin/face-cleanup for the mode chooser. Pure relocation of the
ten-file dashboard cluster via git mv - no behaviour change - plus a
temporary 307 redirect at the old path so this lands without a dead
entry point. The chooser replaces that redirect next.

The navbar is deliberately NOT repointed: NavbarItem highlights via
pathname.startsWith(href), so pointing at /admin/face-cleanup keeps it
active across /scan, /people and the chooser. Repointing it would be
the regression, not the fix.

Four e2e assertions broke on the move. Three are repoints; Playwright
globs match the whole URL, so '**/admin/face-cleanup' does not match
/scan. The fourth needed a real fix rather than a repoint: the
post-apply drain check asserts a person's name is ABSENT, which passes
vacuously on any page that never lists people - or on one that simply
has not finished loading. It now waits for the Eligible faces stat,
which renders unconditionally once a scan is completed and is therefore
independent of the drained-vs-not condition under test.

Tests: 117/117 across the nine guided specs from their new locations,
tsc clean.
