+++ web/src/routes/admin/face-cleanup/scan/+page.svelte
