+++ web/src/routes/admin/face-cleanup/scan/AdvancedScanModal.svelte
