+++ web/src/routes/admin/face-cleanup/scan/FaceCleanupTable.svelte
