+++ web/src/routes/admin/face-cleanup/scan/ScanChecklist.svelte
